[#82] - Migrate members screen
