+++ js/components/Screens/Modals/InvitePlanMembers/InvitePlanMembersScreen.tsx
@@ -19,7 +19,6 @@ import {
 import { connect, ConnectedProps } from 'react-redux'
 import { useAppTheme } from '../../../../hooks/useAppTheme'
 import usePrevious from '../../../../hooks/usePrevious'
-import { plansSlice } from '../../../../reducers/plans'
 import { usersSlice } from '../../../../reducers/users'
 import { RootState } from '../../../../redux/store'
 import { ModalsNavigatorPropsT } from '../../../../types/ModalsNavigatorProps'
@@ -46,7 +45,7 @@ const connector = connect(
   {
     search: usersSlice.actions.search,
     setSearchResults: usersSlice.actions.setSearchResults,
-    updatePlanMembers: plansSlice.actions.updateMembersForPlanId
+    updatePlanMembers: usersSlice.actions.updatePlanMembers
   }
 )
