+++ js/reducers/plans.ts
@@ -1,16 +1,9 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { PlanDataT, PlanT } from '../types/Plan'
-import { UserT } from '../types/User'
 
-type StateT = {
-  data: Record<string, PlanT | undefined>
-  membersForPlanId: Record<string, UserT[] | undefined>
-}
+type StateT = Record<string, PlanT | undefined>
 
-const initialState: StateT = {
-  data: {},
-  membersForPlanId: {}
-}
+const initialState: StateT = {}
 
 export const plansSlice = createSlice({
   name: 'plans',
@@ -18,46 +11,21 @@ export const plansSlice = createSlice({
   reducers: {
     fetch: (_, __: PayloadAction<{ loader: string }>) => {},
     onFetch: (state, action: PayloadAction<Record<string, PlanT>>) => {
-      state.data = action.payload
+      state = action.payload
     },
     create: (_, __: PayloadAction<PlanDataT>) => {},
     onCreate: (state, action: PayloadAction<PlanT>) => {
-      state.data[action.payload.id] = action.payload
-    },
-    fetchMembersForPlanId: (
-      _,
-      __: PayloadAction<{ planId: string; loader?: string }>
-    ) => {},
-    updateMembersForPlanId: (
-      _,
-      __: PayloadAction<{ planId: string; newMembers: UserT[] }>
-    ) => {},
-    setMembersForPlanId: (
-      state,
-      action: PayloadAction<{ planId: string; members: UserT[] }>
-    ) => {
-      const plan = state.data[action.payload.planId]
-
-      if (!plan) {
-        return
-      }
-
-      state.data[action.payload.planId] = {
-        ...plan,
-        userIds: action.payload.members.map(m => m.id)
-      }
-      state.membersForPlanId[action.payload.planId] = action.payload.members
+      state[action.payload.id] = action.payload
     },
     clear: (
       state,
       action: PayloadAction<{ clearRelatedDataOnly: boolean }>
     ) => {
+      // TODO: 82 - return just initial state
+
       if (!action.payload.clearRelatedDataOnly) {
         return initialState
       }
-
-      // clear only plan-related data
-      state.membersForPlanId = initialState.membersForPlanId
     }
   }
 })
